requirements.txt
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==6.2.1
pkg-resources==0.0.0
psycopg2==2.8.4
pur==5.2.2
py-bcrypt==0.4
python-dateutil==2.8.0
six==1.12.0
uWSGI==2.0.18
Werkzeug==0.16.0
WTForms==2.2.1
